+++ RELEASE-NOTES.md
@@ -12,6 +12,7 @@
 * Added `create` command to `spcs image-repository`.
 ## Fixes and improvements
 * Restricted permissions of automatically created files
+* Fixed bug where `spcs service create` would not throw error if service with specified name already exists.
 
 
 # v2.0.0